Split build command validation tests
