+++ tests/integration/cli_build/build_command_validation/target_metadata.rs
@@ -1,0 +1,98 @@
+use std::process::Command;
+
+#[test]
+fn build_command_build_zen_rejects_duplicate_target_fields() {
+    assert_build_command_rejects_target_metadata(
+        r#"
+build = (b: Builder) Result<BuildConfig, BuildError> {
+    b.add(Executable {
+        name: "app",
+        name: "tool",
+        main: "app.zen",
+        out_dir: "build/app/",
+    })
+    .Ok(b.config())
+}
+"#,
+        "duplicate field `name` in `Executable` build target",
+    );
+}
+
+#[test]
+fn build_command_build_zen_rejects_missing_required_target_fields() {
+    assert_build_command_rejects_target_metadata(
+        r#"
+build = (b: Builder) Result<BuildConfig, BuildError> {
+    b.add(Executable {
+        name: "app",
+        main: "app.zen",
+    })
+    .Ok(b.config())
+}
+"#,
+        "missing required field `out_dir` in `Executable` build target",
+    );
+}
+
+#[test]
+fn build_command_build_zen_rejects_invalid_target_field_types() {
+    assert_build_command_rejects_target_metadata(
+        r#"
+build = (b: Builder) Result<BuildConfig, BuildError> {
+    b.add(Executable {
+        name: "app",
+        main: "app.zen",
+        out_dir: 42,
+    })
+    .Ok(b.config())
+}
+"#,
+        "field `out_dir` in `Executable` build target must be a string",
+    );
+}
+
+#[test]
+fn build_command_build_zen_rejects_dynamic_target_adds_before_execution() {
+    assert_build_command_rejects_target_metadata(
+        r#"
+build = (b: Builder) Result<BuildConfig, BuildError> {
+    enabled = true
+    enabled ?
+        | true {
+            b.add(Executable { name: "app", main: "app.zen", out_dir: "build/app/" })
+        }
+        | false {
+            b.add(Executable { name: "tool", main: "tool.zen", out_dir: "build/tool/" })
+        }
+    .Ok(b.config())
+}
+"#,
+        "build targets must be added in the deterministic build graph body",
+    );
+}
+
+fn assert_build_command_rejects_target_metadata(build_source: &str, expected_diagnostic: &str) {
+    let tmp = tempfile::tempdir().expect("create temp dir");
+    std::fs::write(tmp.path().join("build.zen"), build_source).expect("write build.zen");
+    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
+        .args(["build", "build.zen"])
+        .current_dir(tmp.path())
+        .output()
+        .expect("run zen build build.zen");
+
+    assert!(
+        !output.status.success(),
+        "zen build build.zen unexpectedly succeeded: stdout={}, stderr={}",
+        String::from_utf8_lossy(&output.stdout),
+        String::from_utf8_lossy(&output.stderr)
+    );
+    let stderr = String::from_utf8_lossy(&output.stderr);
+    assert!(
+        stderr.contains(expected_diagnostic),
+        "expected target metadata diagnostic `{expected_diagnostic}`, stderr={stderr}"
+    );
+    assert!(
+        !tmp.path().join("build").exists(),
+        "build command should not create outputs after target metadata validation fails"
+    );
+}
